seperate concerns in views
