+++ td_biblio/views/__init__.py
@@ -1,0 +1,9 @@
+# coding: utf-8
+__author__ = 'Alison Mukoma <mukomalison@gmail.com>'
+__license__ = 'MIT'
+__copyright__ = 'tailordev.fr'
+
+from .entry_batch_import import * # noqa
+from .entry_list import * # noqa
+from .find_duplicated_authors import * # noqa
+from .mixins import * # noqa
